services/producer/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
click==8.2.1
colorama==0.4.6
confluent-kafka==2.11.1
dotenv==0.9.9
fastapi==0.116.1
h11==0.16.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
